archive/lh360/pyproject.toml
[project]
name = "lh360"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.96.0",
    "fastapi>=0.115.0",
    "google-api-python-client>=2.194.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "httpx>=0.28.1",
    "mcp>=1.27.0",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "sse-starlette>=2.1.0",
    "uvicorn>=0.32.0",
]
